chore: remove addNote method from edgeless block
