+++ packages/blocks/src/root-block/edgeless/gfx-tool/note-tool.ts
@@ -1,140 +1,20 @@
 import type { SurfaceBlockComponent } from '@blocksuite/affine-block-surface';
 import type { PointerEventState } from '@blocksuite/block-std';
 
-import { focusTextModel } from '@blocksuite/affine-components/rich-text';
-import { type NoteBlockModel, NoteDisplayMode } from '@blocksuite/affine-model';
-import {
-  EditPropsStore,
-  TelemetryProvider,
-} from '@blocksuite/affine-shared/services';
-import { handleNativeRangeAtPoint } from '@blocksuite/affine-shared/utils';
-import { BaseTool, type GfxController } from '@blocksuite/block-std/gfx';
-import { type IPoint, Point, serializeXYWH } from '@blocksuite/global/utils';
+import { EditPropsStore } from '@blocksuite/affine-shared/services';
+import { BaseTool } from '@blocksuite/block-std/gfx';
+import { Point } from '@blocksuite/global/utils';
 import { effect } from '@preact/signals-core';
 
 import {
   hasClassNameInList,
   type NoteChildrenFlavour,
 } from '../../../_common/utils/index.js';
+import { addNote } from '../utils/common.js';
 import {
-  DEFAULT_NOTE_HEIGHT,
-  DEFAULT_NOTE_OFFSET_X,
-  DEFAULT_NOTE_OFFSET_Y,
-  DEFAULT_NOTE_WIDTH,
   EXCLUDING_MOUSE_OUT_CLASS_LIST,
-  NOTE_MIN_HEIGHT,
 } from '../utils/consts.js';
-import { addBlock } from '../utils/crud.js';
 import { DraggingNoteOverlay, NoteOverlay } from '../utils/tool-overlay.js';
-
-type NoteOptions = {
-  childFlavour: NoteChildrenFlavour;
-  childType: string | null;
-  collapse: boolean;
-};
-
-function addNoteWithPoint(
-  gfx: GfxController,
-  point: IPoint,
-  options: {
-    width?: number;
-    height?: number;
-    parentId?: string;
-    noteIndex?: number;
-    offsetX?: number;
-    offsetY?: number;
-    scale?: number;
-  } = {}
-) {
-  const {
-    width = DEFAULT_NOTE_WIDTH,
-    height = DEFAULT_NOTE_HEIGHT,
-    offsetX = DEFAULT_NOTE_OFFSET_X,
-    offsetY = DEFAULT_NOTE_OFFSET_Y,
-    parentId = gfx.doc.root?.id,
-    noteIndex: noteIndex,
-    scale = 1,
-  } = options;
-  const [x, y] = gfx.viewport.toModelCoord(point.x, point.y);
-  const blockId = addBlock(
-    gfx.std,
-    'affine:note',
-    {
-      xywh: serializeXYWH(
-        x - offsetX * scale,
-        y - offsetY * scale,
-        width,
-        height
-      ),
-      displayMode: NoteDisplayMode.EdgelessOnly,
-    },
-    parentId,
-    noteIndex
-  );
-
-  gfx.std.getOptional(TelemetryProvider)?.track('CanvasElementAdded', {
-    control: 'canvas:draw',
-    page: 'whiteboard editor',
-    module: 'toolbar',
-    segment: 'toolbar',
-    type: 'note',
-  });
-
-  return blockId;
-}
-
-function addNote(
-  gfx: GfxController,
-  point: Point,
-  options: NoteOptions,
-  width = DEFAULT_NOTE_WIDTH,
-  height = DEFAULT_NOTE_HEIGHT
-) {
-  const noteId = addNoteWithPoint(gfx, point, {
-    width,
-    height,
-  });
-
-  const doc = gfx.doc;
-
-  const blockId = doc.addBlock(
-    options.childFlavour,
-    { type: options.childType },
-    noteId
-  );
-  if (options.collapse && height > NOTE_MIN_HEIGHT) {
-    const note = doc.getBlockById(noteId) as NoteBlockModel;
-    doc.updateBlock(note, () => {
-      note.edgeless.collapse = true;
-      note.edgeless.collapsedHeight = height;
-    });
-  }
-  gfx.tool.setTool('default');
-
-  // Wait for edgelessTool updated
-  requestAnimationFrame(() => {
-    const blocks =
-      (doc.root?.children.filter(
-        child => child.flavour === 'affine:note'
-      ) as BlockSuite.EdgelessBlockModelType[]) ?? [];
-    const element = blocks.find(b => b.id === noteId);
-    if (element) {
-      gfx.selection.set({
-        elements: [element.id],
-        editing: true,
-      });
-
-      // Waiting dom updated, `note mask` is removed
-      if (blockId) {
-        focusTextModel(gfx.std, blockId);
-      } else {
-        // Cannot reuse `handleNativeRangeClick` directly here,
-        // since `retargetClick` will re-target to pervious editor
-        handleNativeRangeAtPoint(point.x, point.y);
-      }
-    }
-  });
-}
 
 export type NoteToolOption = {
   childFlavour: NoteChildrenFlavour;
@@ -228,7 +108,7 @@ export class NoteTool extends BaseTool<NoteToolOption> {
       collapse: false,
     };
     const point = new Point(e.point.x, e.point.y);
-    addNote(this.gfx, point, options);
+    addNote(this.std, point, options);
   }
 
   override deactivate() {
@@ -257,7 +137,7 @@ export class NoteTool extends BaseTool<NoteToolOption> {
     this.doc.captureSync();
 
     addNote(
-      this.gfx,
+      this.std,
       point,
       options,
       Math.max(width, DEFAULT_NOTE_WIDTH),
